queen check updated to include opponent + own
